crimson/os/seastore: refactor LBAMapping
So that PhyscalNodeMapping doesn't contain LBAMapping specific
interfaces
